I have been too busy to do much work on this, which is quite unfortunate. Luckily I have made some changes to my work schedule, so starting around the end of this upcoming week I will be working quite a bit less and will hopefully be able to begin making large chunks of progress again.
I did a bit of debugging what I had here to see why it wasn't working, and I have discovered that I still have learning to do regarding the use of VAOs. What I have right now in my SimpleVertexBuffer for the VAO is not correct. I probably may want to consider moving my VAO to be in it's own class. Also, a VAO is not valid until it is bound.

Oh well...   This is as far as I am getting today because I have to go to work. Hopefuly my commit frequency will increase starting sometime within the next 7 days, and I will be able to actually get something working. The untapped potential in what I know I can create here is driving me sorta crazy, I just gotta not think at all and then everything is fine. Which is unfortunate. I guess a plus to all of this is I am being made very aware of how fortunate I am to have time to spend to work on this stuff.