It's been a while since I have had a chance to do much of anything of worth for commiting... Most recently I had been working on the CallbackEventStream buffer class, but the design for that is not good for what its intended use is, and it doesn't work because for some reason it indexes out of the valid range of it's member-field vectors, and on top of that, it doesn't perform synchronization correctly between the callbacks and application. The plan from here is to read up more on C++11's std::condition_variable, C++14's std::shared_timed_mutex, and C++17's std::shared_mutex since from the brief reading I have done they seem to be more along the lines of the synchronization primitive required.
Another recent development has been my increased understanding of what a Context Reset is (thanks to the book OpenGL SuperBible 7e ch. 15). I have begun to plan out a general strategy for dealing with context-reset events, since they are both unpredicatble and catatrosphic if no measures are in place for recovery. There is currently a long blurb explaining the basic premise in the file RenderDemoBase.h.

Based off some research into GLFW's documentation, a context reset will require a complete recreation of the GLFW window and context.

Thus the plan going forward from here: Rebuild the CallbackEventStream class and build that into a larger WindowAttachment class for attaching to GLFW with the window user pointer slot, which will also involve rewriting and improving the GLFW_Init class. This rewrite will also emphasis a robust context-reset-recovery strategy, thus improving the foundation of the entire application all at once.
This is going to be a lot of work, but it will be very worth it in the end.