Based off some quick tests I just did with the GLFW window callback functions, it most definitly appears that the callbacks are:   i) executed by the main thread,  ii) they are have a predictable order in which they will execute, and perhaps most importantly iii) They do not execute at the same time.
So all this preparation and thought about how to handle asynchronus callbacks was not necessary. Luckily I haven't gotten very far with my assynchronus implementation. I'm also a bit disappointed because I was hoping to use this as an exercise in increasing my proficiency with  synchronization primitives.

Oh well, at least this greatly simplifies the task I was trying to accomplish...