2 Minor Updates: i) Moved the dependency on the header 'RelativeFilepathsToResources.h' from the AssetLoadingDemo header file to this here implementation file.
ii) Changed several places where I originally had declared variables as  'size_t' under the assumption that this would always match 'GLsizei', which in hindsight is a very bad idea. I now understand all too well that the underlying type of 'size_t' is dependent upon the underlying architecture. This marks the start of my plan to slowly begin replacing every single declaration of size_t I come across with a more specific type (will depend on the context). Until this is completed, the Application will be incompatible with running in a 64-bit architecture.