The member function 'bool checkKeyboardInput()'  was missing a return statement at it's end, which results in undefined behavior. By luck, this undefined behavior never seemed to cause any issues and had thus slipped by unnoticed by me for nearly 2 years. Until now that is. I added a return statement to this function.
Also, the constant variable representing the relative filepath to the shader code files has been renamed so as to increase the codes clarity.

Also, changed a few local variables previously declared as 'size_t' to now be 'uint64_t'. The advantage to using 'uint64_t' is that this variable will now always have the same width across various platforms. I still have a wide-spread issue of variables of type 'size_t' being used throughout the project. The issue with 'size_t' is it's width is platform dependent