Added the 3 custom shader parameters as uniforms to these shaders to open them up for more variety. This is currently all just experimental stuff and what they are used for now doesn't really make any sense. Having them present just adds more variety to affect what the shader is doing through manipulating the values of these variables.
I'm committing both the frag and the vert shader in tandem because they both are designed to be used together. Please note that at this time, only the vert shader does anything with these custom shader parameters. The frag shader only declares them and never uses them. This will probably change in the future